Can successfully get account info with retrofit
